+++ eq-author/src/features/design/entities/QuestionPage/Design/enhancers/answer/withDeleteAnswer/deleteAnswer.graphql
