+++ eq-author/src/features/design/entities/QuestionPage/Design/enhancers/answer/withUpdateAnswer/index.test.js
@@ -1,4 +1,4 @@
-import { mapMutateToProps } from "./withUpdateAnswer";
+import { mapMutateToProps } from ".";
 
 describe("withUpdateAnswer", () => {
   let mutate;
