+++ eq-author/src/features/design/entities/QuestionPage/Design/enhancers/answer/withUpdateAnswer/updateAnswer.graphql
@@ -1,4 +1,4 @@
-#import "./fragments/answer.graphql"
+#import "graphql/fragments/answer.graphql"
 
 mutation UpdateAnswer($input: UpdateAnswerInput!) {
   updateAnswer(input: $input) {
