+++ eq-author/src/features/design/entities/QuestionPage/Design/enhancers/option/withDeleteOption.js
