+++ eq-author/src/features/design/entities/QuestionPage/Design/enhancers/option/withDeleteOption.test.js
